feat(aggretated-metrics): ignore agg metrics from limits
* ignore aggregated metric streams from limits
