+++ pkg/distributor/distributor.go
@@ -1,7 +1,6 @@
 package distributor
 
 import (
-	"bytes"
 	"context"
 	"flag"
 	"fmt"
@@ -11,15 +10,11 @@ import (
 	"strconv"
 	"strings"
 	"time"
-	"unicode"
-	"unsafe"
 
-	"github.com/buger/jsonparser"
 	"github.com/go-kit/log"
 	"github.com/go-kit/log/level"
 	"github.com/gogo/status"
 	"github.com/prometheus/prometheus/model/labels"
-	"go.opentelemetry.io/collector/pdata/plog"
 	"google.golang.org/grpc/codes"
 
 	"github.com/grafana/dskit/httpgrpc"
@@ -39,6 +34,7 @@ import (
 
 	"github.com/grafana/loki/v3/pkg/analytics"
 	"github.com/grafana/loki/v3/pkg/compactor/retention"
+	"github.com/grafana/loki/v3/pkg/detection"
 	"github.com/grafana/loki/v3/pkg/distributor/clientpool"
 	"github.com/grafana/loki/v3/pkg/distributor/shardstreams"
 	"github.com/grafana/loki/v3/pkg/distributor/writefailures"
@@ -46,7 +42,6 @@ import (
 	"github.com/grafana/loki/v3/pkg/ingester/client"
 	"github.com/grafana/loki/v3/pkg/loghttp/push"
 	"github.com/grafana/loki/v3/pkg/logproto"
-	"github.com/grafana/loki/v3/pkg/logql/log/logfmt"
 	"github.com/grafana/loki/v3/pkg/logql/syntax"
 	"github.com/grafana/loki/v3/pkg/runtime"
 	"github.com/grafana/loki/v3/pkg/util"
@@ -63,28 +58,12 @@ const (
 
 	LabelServiceName = "service_name"
 	ServiceUnknown   = "unknown_service"
-	LevelLabel       = "detected_level"
-	LogLevelUnknown  = "unknown"
-
-	logLevelDebug    = "debug"
-	logLevelInfo     = "info"
-	logLevelWarn     = "warn"
-	logLevelError    = "error"
-	logLevelFatal    = "fatal"
-	logLevelCritical = "critical"
-	logLevelTrace    = "trace"
 )
 
 var (
 	maxLabelCacheSize = 100000
 	rfStats           = analytics.NewInt("distributor_replication_factor")
 )
-
-var allowedLabelsForLevel = map[string]struct{}{
-	"level": {}, "LEVEL": {}, "Level": {},
-	"severity": {}, "SEVERITY": {}, "Severity": {},
-	"lvl": {}, "LVL": {}, "Lvl": {},
-}
 
 // Config for a Distributor.
 type Config struct {
@@ -407,11 +386,11 @@ func (d *Distributor) Push(ctx context.Context, req *logproto.PushRequest) (*log
 					} else if levelFromMetadata, ok := hasAnyLevelLabels(structuredMetadata); ok {
 						logLevel = levelFromMetadata
 					} else {
-						logLevel = DetectLogLevelFromLogEntry(entry, structuredMetadata)
+						logLevel = detection.DetectLogLevelFromLogEntry(entry, structuredMetadata)
 					}
-					if logLevel != LogLevelUnknown && logLevel != "" {
+					if logLevel != detection.LogLevelUnknown && logLevel != "" {
 						entry.StructuredMetadata = append(entry.StructuredMetadata, logproto.LabelAdapter{
-							Name:  LevelLabel,
+							Name:  detection.LevelLabel,
 							Value: logLevel,
 						})
 					}
@@ -562,7 +541,7 @@ func (d *Distributor) Push(ctx context.Context, req *logproto.PushRequest) (*log
 }
 
 func hasAnyLevelLabels(l labels.Labels) (string, bool) {
-	for lbl := range allowedLabelsForLevel {
+	for lbl := range detection.AllowedLabelsForLevel {
 		if l.Has(lbl) {
 			return l.Get(lbl), true
 		}
@@ -787,7 +766,8 @@ func (d *Distributor) parseStreamLabels(vContext validationContext, key string, 
 	}
 
 	// We do not want to count service_name added by us in the stream limit so adding it after validating original labels.
-	if !ls.Has(LabelServiceName) && len(vContext.discoverServiceName) > 0 {
+	if !ls.Has(LabelServiceName) && !ls.Has(detection.AggregatedMetricLabel) &&
+		len(vContext.discoverServiceName) > 0 {
 		serviceName := ServiceUnknown
 		for _, labelName := range vContext.discoverServiceName {
 			if labelVal := ls.Get(labelName); labelVal != "" {
@@ -888,130 +868,4 @@ func newRingAndLifecycler(cfg RingConfig, instanceCount *atomic.Uint32, logger l
 // distributor. $EFFECTIVE_RATE_LIMIT = $GLOBAL_RATE_LIMIT / $NUM_INSTANCES.
 func (d *Distributor) HealthyInstancesCount() int {
 	return int(d.healthyInstancesCount.Load())
-}
-
-func DetectLogLevelFromLogEntry(entry logproto.Entry, structuredMetadata labels.Labels) string {
-	// otlp logs have a severity number, using which we are defining the log levels.
-	// Significance of severity number is explained in otel docs here https://opentelemetry.io/docs/specs/otel/logs/data-model/#field-severitynumber
-	if otlpSeverityNumberTxt := structuredMetadata.Get(push.OTLPSeverityNumber); otlpSeverityNumberTxt != "" {
-		otlpSeverityNumber, err := strconv.Atoi(otlpSeverityNumberTxt)
-		if err != nil {
-			return logLevelInfo
-		}
-		if otlpSeverityNumber == int(plog.SeverityNumberUnspecified) {
-			return LogLevelUnknown
-		} else if otlpSeverityNumber <= int(plog.SeverityNumberTrace4) {
-			return logLevelTrace
-		} else if otlpSeverityNumber <= int(plog.SeverityNumberDebug4) {
-			return logLevelDebug
-		} else if otlpSeverityNumber <= int(plog.SeverityNumberInfo4) {
-			return logLevelInfo
-		} else if otlpSeverityNumber <= int(plog.SeverityNumberWarn4) {
-			return logLevelWarn
-		} else if otlpSeverityNumber <= int(plog.SeverityNumberError4) {
-			return logLevelError
-		} else if otlpSeverityNumber <= int(plog.SeverityNumberFatal4) {
-			return logLevelFatal
-		}
-		return LogLevelUnknown
-	}
-
-	return extractLogLevelFromLogLine(entry.Line)
-}
-
-func extractLogLevelFromLogLine(log string) string {
-	logSlice := unsafe.Slice(unsafe.StringData(log), len(log))
-	var v []byte
-	if isJSON(log) {
-		v = getValueUsingJSONParser(logSlice)
-	} else {
-		v = getValueUsingLogfmtParser(logSlice)
-	}
-
-	switch {
-	case bytes.EqualFold(v, []byte("trace")), bytes.EqualFold(v, []byte("trc")):
-		return logLevelTrace
-	case bytes.EqualFold(v, []byte("debug")), bytes.EqualFold(v, []byte("dbg")):
-		return logLevelDebug
-	case bytes.EqualFold(v, []byte("info")), bytes.EqualFold(v, []byte("inf")):
-		return logLevelInfo
-	case bytes.EqualFold(v, []byte("warn")), bytes.EqualFold(v, []byte("wrn")):
-		return logLevelWarn
-	case bytes.EqualFold(v, []byte("error")), bytes.EqualFold(v, []byte("err")):
-		return logLevelError
-	case bytes.EqualFold(v, []byte("critical")):
-		return logLevelCritical
-	case bytes.EqualFold(v, []byte("fatal")):
-		return logLevelFatal
-	default:
-		return detectLevelFromLogLine(log)
-	}
-}
-
-func getValueUsingLogfmtParser(line []byte) []byte {
-	equalIndex := bytes.Index(line, []byte("="))
-	if len(line) == 0 || equalIndex == -1 {
-		return nil
-	}
-
-	d := logfmt.NewDecoder(line)
-	for !d.EOL() && d.ScanKeyval() {
-		if _, ok := allowedLabelsForLevel[string(d.Key())]; ok {
-			return (d.Value())
-		}
-	}
-	return nil
-}
-
-func getValueUsingJSONParser(log []byte) []byte {
-	for allowedLabel := range allowedLabelsForLevel {
-		l, _, _, err := jsonparser.Get(log, allowedLabel)
-		if err == nil {
-			return l
-		}
-	}
-	return nil
-}
-
-func isJSON(line string) bool {
-	var firstNonSpaceChar rune
-	for _, char := range line {
-		if !unicode.IsSpace(char) {
-			firstNonSpaceChar = char
-			break
-		}
-	}
-
-	var lastNonSpaceChar rune
-	for i := len(line) - 1; i >= 0; i-- {
-		char := rune(line[i])
-		if !unicode.IsSpace(char) {
-			lastNonSpaceChar = char
-			break
-		}
-	}
-
-	return firstNonSpaceChar == '{' && lastNonSpaceChar == '}'
-}
-
-func detectLevelFromLogLine(log string) string {
-	if strings.Contains(log, "info:") || strings.Contains(log, "INFO:") ||
-		strings.Contains(log, "info") || strings.Contains(log, "INFO") {
-		return logLevelInfo
-	}
-	if strings.Contains(log, "err:") || strings.Contains(log, "ERR:") ||
-		strings.Contains(log, "error") || strings.Contains(log, "ERROR") {
-		return logLevelError
-	}
-	if strings.Contains(log, "warn:") || strings.Contains(log, "WARN:") ||
-		strings.Contains(log, "warning") || strings.Contains(log, "WARNING") {
-		return logLevelWarn
-	}
-	if strings.Contains(log, "CRITICAL:") || strings.Contains(log, "critical:") {
-		return logLevelCritical
-	}
-	if strings.Contains(log, "debug:") || strings.Contains(log, "DEBUG:") {
-		return logLevelDebug
-	}
-	return LogLevelUnknown
 }
