+++ pkg/ingester/stream.go
@@ -8,6 +8,7 @@ import (
 	"sync"
 	"time"
 
+	"github.com/grafana/loki/v3/pkg/detection"
 	"github.com/grafana/loki/v3/pkg/runtime"
 
 	"github.com/go-kit/log/level"
@@ -82,6 +83,8 @@ type stream struct {
 	chunkHeadBlockFormat chunkenc.HeadBlockFmt
 
 	configs *runtime.TenantConfigs
+
+	isAggregatedMetric bool
 }
 
 type chunkDesc struct {
@@ -132,7 +135,8 @@ func newStream(
 		chunkFormat:          chunkFormat,
 		chunkHeadBlockFormat: headBlockFmt,
 
-		configs: configs,
+		configs:            configs,
+		isAggregatedMetric: labels.Has(detection.AggregatedMetricLabel),
 	}
 }
